refactor: FSD 설계기준에 맞게 수정
